JDP240201-14_Cart_Entity - implementation

+++ src/main/java/com/kodilla/ecommercee/domain/CartDto.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Setter
 @AllArgsConstructor
 public class CartDto {
-    private int id;
-    private int userId;
+    private Long id;
+    private Long userId;
     private List<Integer> productsIds;
 }
